+++ lib/src/builder.dart
@@ -19,26 +19,23 @@ class _NationsBuilderState extends State<NationsBuilder> {
   @override
   Widget build(BuildContext context) {
     _forceRebuild(context);
-
     return widget.builder(context);
   }
 
   @override
   void initState() {
-    Nations.addListener(() {
-      _forceRebuild(context);
-    });
+    Nations.addListener(() => _forceRebuild(context));
     super.initState();
   }
 
-  void _forceRebuild(BuildContext context) {
-    void rebuild(Element element) {
-      if (element.debugDoingBuild || element.dirty) return;
-      element
-        ..markNeedsBuild()
-        ..visitChildren(rebuild);
-    }
-
-    (context as Element).visitChildren(rebuild);
+  void _rebuild(Element element) {
+    if (element.debugDoingBuild || element.dirty) return;
+    element.markNeedsBuild();
+    element.visitChildren(_rebuild);
   }
+
+  void _forceRebuild(
+    BuildContext context,
+  ) =>
+      (context as Element).visitChildren(_rebuild);
 }