+++ lib/src/extract/base.dart
@@ -1,7 +1,6 @@
 import 'package:flutter_readable/flutter_readable.dart';
 import 'package:nations/nations.dart';
 import 'package:nations/src/helpers/extractors.dart';
-import 'package:nations/src/helpers/validation.dart';
 part 'string.dart';
 part 'map.dart';
 part 'not_found.dart';