+++ lib/src/extract/map.dart
@@ -10,14 +10,12 @@ class MapData extends ExtractedData<Map<String, Object?>> {
 
   @override
   String? toGender([Gender? gender]) {
-    if (isValidGenderMap(data)) {
-      if (gender == null) {
-        return toGender(Nations.config.defaultGender);
-      } else if (gender == Gender.male && data['male'] is String) {
-        return data['male']! as String;
-      } else if (gender == Gender.female && data['female'] is String) {
-        return data['female']! as String;
-      }
+    if (gender == null) {
+      return toGender(Nations.config.defaultGender);
+    } else if (gender == Gender.male && data['male'] is String) {
+      return data['male']! as String;
+    } else if (gender == Gender.female && data['female'] is String) {
+      return data['female']! as String;
     }
   }