+++ lib/src/helpers/files.dart
@@ -3,10 +3,10 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-/// *  load translation files from the `nations` package base on the locale
-/// * if the locale doesn't exist in your assets it will try to get
-/// *  the file  named with the language code
-/// * if also is empty it will loads the fallback locale
+// / *  load translation files from the `nations` package base on the locale
+// / * if the locale doesn't exist in your assets it will try to get
+// / *  the file  named with the language code
+// / * if also is empty it will loads the fallback locale
 // Future<Map<String, dynamic>?> loadPackageTranslation({
 //   required Locale locale,
 //   required String packageName,